Some manual merging required

+++ matlab/loadLbl.m
@@ -1,4 +1,8 @@
-function lbl = loadLbl(lbl_file)
+function lbl = loadLbl(lbl_file,varargin)
+
+keeptempdir = myparse(varargin,...
+  'keeptempdir',false ...
+  );
 
 tname = tempname;
 mkdir(tname);
@@ -18,9 +22,11 @@ end
 
 lbl = load(rawLblFile,'-mat');
 
-[success, message, ~] = rmdir(tname,'s');
-if ~success
-  error('Could not clear the temp directory %s\n',message);
-else
-  fprintf('Cleared out temp directory %s\n',tname);
+if ~keeptempdir
+  [success, message, ~] = rmdir(tname,'s');
+  if ~success
+    error('Could not clear the temp directory %s\n',message);
+  else
+    fprintf('Cleared out temp directory %s\n',tname);
+  end
 end
